+++ src/bandcamp/content_module.js
@@ -1,18 +1,29 @@
 import { Release } from "../app/release.js";
-import { contentChangeWithPolling, createDatalistFromArray, createElementFromHTML, input, selectElementWithContent, setDataAttribute } from "../modules/html.js";
-import { containsOneOf, splitString, injectCSSFile, injectJSFile, isEmptyArray, countOccurrences, removeBrackets } from "../modules/utils.js";
+import { click, contentChangeWithPolling, createDatalistFromArray, createElementFromHTML, input, isElementDisplayNone, isHtmlElement, selectElementWithContent, setDataAttribute } from "../modules/html.js";
+import { getAlbumRelease } from "../modules/schema.js";
+import { addReleaseHistory, findReleaseInStorage, saveRelease } from "../modules/storage.js";
+import { containsOneOf, splitString, injectCSSFile, injectJSFile, isEmptyArray, countOccurrences, removeBrackets, isObject } from "../modules/utils.js";
 import { PageType, PageTypeDetector } from "./bandcamp.js";
 import { getBandPhotoSrc, getReleasesData } from "./html.js";
 
 export function main () {
   console.log('B2D: CONTENT AS MODULE');
   const pageType = (new PageTypeDetector()).detect();
-
+  setupReleaseCollectedByWidget(pageType);
   setupIsotope(pageType);
   setupSendMessageToPopup(pageType);
   setupBCDataEventListener(pageType);
   injectJSFile(chrome.runtime.getURL('src/bandcamp/script.js'));
   injectCSSFile(chrome.runtime.getURL('src/bandcamp/styles.css'));
+}
+
+function getMusicAlbumSchemaData() {
+  const jsonLdScript = document.querySelector('script[type="application/ld+json"]');
+  return jsonLdScript ? JSON.parse(jsonLdScript.textContent) : null;
+}
+
+function getCurrentUrl() {
+  return window.location.href;
 }
 
 /**
@@ -27,28 +38,18 @@ function setupBCDataEventListener(pageType) {
   window.addEventListener('BC_Data', (e) => {
     // Getting data from script.js
     const {TralbumData, BandData} = e.detail;
-    const currentTabUrl = window.location.href;
-    const storage = chrome.storage.local;
 
-    // storage.clear();
-
-    storage.get([currentTabUrl], (result) => {
+    findReleaseInStorage(getCurrentUrl(), null, (url) => {
       // Save release data to the storage if it doesn't exist
-      if (!result[currentTabUrl] || !result[currentTabUrl]['release']) {
-        const { schemaData, coverSrc } = extractReleaseData();
-        const release = Release.fromBandcampData(
-          TralbumData,
-          BandData,
-          schemaData,
-          coverSrc
-        );
+      const { schemaData, coverSrc } = extractReleaseData();
+      const release = Release.fromBandcampData(
+        TralbumData,
+        BandData,
+        schemaData,
+        coverSrc
+      );
 
-        storage.set({ [currentTabUrl]: { release: release.toJSON() } }, () => {
-          console.log("B2D: Release data was saved in local storage");
-        });
-      } else {
-        console.log("B2D: Release data already exists");
-      }
+      saveRelease(url, release);
     });
   });
 }
@@ -104,6 +105,76 @@ function setupSendMessageToPopup(pageType) {
 function getArtistFilterValue() {
   const artistFilter = document.querySelector('#b2dArtistFilter');
   return artistFilter.value;
+}
+
+/**
+ * @param {PageType} pageType
+ */
+function setupReleaseCollectedByWidget(pageType) {
+  if (!pageType.isAlbum()) return;
+
+  const musicAlbumData = getMusicAlbumSchemaData();
+  const bcCollectedByContainer = document.querySelector('.collected-by');
+  const collectedByMessage = bcCollectedByContainer.querySelector('.message');
+
+  const loadingElement = createElementFromHTML('<span class="loading">Bandcamp To Discogs:<br>Calculating revenue information</span>');
+  const moreWritingElement = bcCollectedByContainer.querySelector('.more-writing');
+  const loadingWritingElement = bcCollectedByContainer.querySelector('.loading-writing');
+  const moreThumbsElement = bcCollectedByContainer.querySelector('.more-thumbs');
+  const loadingThumbsElement = bcCollectedByContainer.querySelector('.loading-thumbs');
+
+  function outputInformation() {
+    const fanElements = bcCollectedByContainer.querySelectorAll('.pic');
+    const fanAmount = fanElements.length;
+    let sponsoredText = `supported by <strong>${fanAmount}</strong> people.`;
+
+    let categoryValues = {};
+    categoryValues.fans = fanAmount;
+
+    const digitalAlbumRelease = getAlbumRelease(musicAlbumData, 'DigitalFormat');
+    const cdAlbumRelease = getAlbumRelease(musicAlbumData, 'CDFormat');
+
+    if (isObject(digitalAlbumRelease)) {
+      const revenue = fanAmount * digitalAlbumRelease.offers.price;
+      sponsoredText += `<br>Digital format potential revenue is ${revenue} ${digitalAlbumRelease.offers.priceCurrency}.`;
+      categoryValues.digitalRevenue = revenue;
+    }
+
+    if (isObject(cdAlbumRelease)) {
+      const revenue = fanAmount * cdAlbumRelease.offers.price;
+      sponsoredText += `<br>CD format potential revenue is ${revenue} ${cdAlbumRelease.offers.priceCurrency}.`;
+      categoryValues.cdRevenue = revenue;
+    }
+
+    collectedByMessage.innerHTML = sponsoredText;
+    addReleaseHistory(getCurrentUrl(), categoryValues);
+  }
+
+  function clickMoreLinks() {
+    const isLoadingThumbs = isHtmlElement(loadingThumbsElement) && !isElementDisplayNone(loadingThumbsElement);
+    const isLoadingWritings = isHtmlElement(loadingWritingElement) && !isElementDisplayNone(loadingWritingElement);
+    const hasMoreWritings = isHtmlElement(moreWritingElement) && !isElementDisplayNone(moreWritingElement);
+    const hasMoreThumbs = isHtmlElement(moreThumbsElement) && !isElementDisplayNone(moreThumbsElement);
+
+    if (!isLoadingWritings && !isLoadingThumbs && !hasMoreWritings && !hasMoreThumbs) {
+      clearInterval(intervalId);
+      outputInformation();
+      return;
+    }
+
+    if (!isLoadingWritings && hasMoreWritings) {
+      click(moreWritingElement);
+    }
+
+    if (!isLoadingThumbs && hasMoreThumbs) {
+      click(moreThumbsElement);
+    }
+  }
+
+  // Show loading indicator
+  collectedByMessage.innerHTML = '';
+  collectedByMessage.append(loadingElement);
+  const intervalId = setInterval(clickMoreLinks, 500);
 }
 
 /**
@@ -197,7 +268,9 @@ function createArtistFilterWidget(releases) {
 
 function createAlbumAmountWidget(releases) {
   return createElementFromHTML(
-`<div class="b2d-widget" id="b2d-albumAmount" title="The amount of releases on the page">Releases: <span class="b2d-viewed">${releases.length}</span> / <strong>${releases.length}</strong></div>`);
+`<div class="b2d-albumAmount b2d-widget" title="The amount of releases on the page">
+Releases: <span class="b2d-visible">${releases.length}</span> / <span class="b2d-total">${releases.length}</span>
+</div>`);
 }
 
 function setupArtistFilterElement(artistFilterElement, iso, albumAmountWidget) {
@@ -211,7 +284,7 @@ function setupArtistFilterElement(artistFilterElement, iso, albumAmountWidget) {
     const filter = selectedValue ? `[data-filter-artist*="${selectedValue}"]` : '*';
     iso.arrange({ filter: filter });
 
-    albumAmountWidget.querySelector('.b2d-viewed').innerHTML = iso.getFilteredItemElements().length;
+    albumAmountWidget.querySelector('.b2d-visible').innerHTML = iso.getFilteredItemElements().length;
 
     // try to updata images
     window.scrollBy(0, 1);