+++ src/modules/storage.js
@@ -1,6 +1,6 @@
 import { Release } from "../app/release.js";
-import { isValidBandcampURL } from "../popup/helpers.js";
-import { isArray } from "./utils.js";
+import { isValidBandcampURL } from "../bandcamp/html.js";
+import { isArray, isFunction, isObject } from "./utils.js";
 
 const storage = chrome.storage.local;
 
@@ -22,18 +22,18 @@ export function findAllReleasesInStorage(onFind) {
       }
     }
 
-    onFind(releases);
+    if (isFunction(onFind)) onFind(releases);
   });
 }
 
-export function findReleaseInStorage(url, onFind) {
+export function findReleaseInStorage(url, onFind, onMissing) {
   storage.get([url], (result) => {
     if (result[url] && result[url]['release']) {
       const release = createReleaseFromStorageItem(result[url]);
-
-      onFind(release);
+      if (isFunction(onFind)) onFind(release);
     } else {
       console.log("B2D: Release data doesn't exists", url);
+      if (isFunction(onMissing)) onMissing(url);
     }
   });
 
@@ -42,7 +42,7 @@ export function findReleaseInStorage(url, onFind) {
 export function findReleasesInStorage(urls, onFind) {
   storage.get(urls, result => {
     let releases = Object.values(result).map(item => createReleaseFromStorageItem(item));
-    onFind(releases);
+    if (isFunction(onFind)) onFind(releases);
   });
 }
 
@@ -50,8 +50,67 @@ export function findMissingKeysInStorage(keys, onFind) {
   storage.get(keys, result => {
     let foundKeys = Object.keys(result);
     let missingKeys = keys.filter(key => !foundKeys.includes(key));
-    onFind(missingKeys);
+    if (isFunction(onFind)) onFind(missingKeys);
   });
+}
+
+/**
+ * @param {String} key
+ * @param {Release} release
+ */
+export function saveRelease(key, release) {
+  storage.set({ [key]: { release: release.toJSON() } }, () => {
+    console.log("B2D: Release data was saved in the local storage");
+  });
+}
+
+/**
+ * @param {String} key
+ * @param {Array} categoryValues
+ */
+export function addReleaseHistory(key, categoryValues) {
+  storage.get([key], (data) => {
+    let releaseData = data[key] || null;
+    if (releaseData === null) return;
+    let history = releaseData.history || {};
+
+    const date = new Date();
+    const dateStr = date.toISOString();
+
+    for (let category in categoryValues) {
+      if (!categoryValues.hasOwnProperty(category)) continue;
+
+      let historyCategory = history[category] || [];
+      if (!isArray(historyCategory)) continue;
+
+      const historyValue = {
+        date: dateStr,
+        value: categoryValues[category]
+      };
+
+      addUniqueValueObjectToArray(historyCategory, historyValue);
+      history[category] = historyCategory;
+    }
+
+    releaseData.history = history;
+
+    storage.set({ [key]: releaseData }, () => {
+      console.log("B2D: Release history for categories was saved successfully", categoryValues);
+    });
+  });
+}
+
+function addUniqueValueObjectToArray(arr, obj) {
+  if (arr.length === 0 || arr[arr.length - 1].value !== obj.value) {
+      arr.push(obj);
+  }
+}
+
+function formatDateYMD(date) {
+  const year = date.getFullYear();
+  const month = (date.getMonth() + 1).toString().padStart(2, '0');
+  const day = date.getDate().toString().padStart(2, '0');
+  return `${year}-${month}-${day}`;
 }
 
 function createReleaseFromStorageItem(storageItem) {
@@ -73,8 +132,6 @@ export function clearStorageByKey(key, onDone) {
         console.error(`Error clearing local storage item with key "${key}": ${chrome.runtime.lastError}`);
       }
 
-      if (typeof onDone === "function") {
-        onDone();
-      }
+      if (isFunction(onDone)) onDone();
     });
 }